extrack result calculation into concern
