+++ app/models/qualification_jump.rb
@@ -14,15 +14,11 @@
 #
 
 class QualificationJump < ApplicationRecord
-  include AcceptsNestedTrack
-
-  SECONDS_BEFORE_START = 10
+  include AcceptsNestedTrack, SubmissionResult
 
   belongs_to :tournament_competitor
   belongs_to :qualification_round
   belongs_to :track, optional: true
-
-  before_save :calculate_result
 
   alias_attribute :competitor, :tournament_competitor
   alias_attribute :round, :qualification_round
@@ -51,35 +47,5 @@ class QualificationJump < ApplicationRecord
 
   def track_comment
     "#{tournament.name} - Qualification #{qualification_round.order}"
-  end
-
-  # TODO!
-  def competitor
-    tournament_competitor
-  end
-
-  def calculate_result
-    return unless track
-    return unless start_time
-    return if result
-
-    intersection_point = PathIntersectionFinder.new(
-      track_points,
-      finish_line
-    ).execute
-
-    self.result = (intersection_point[:gps_time].to_f - start_time.to_f).round(3)
-  end
-
-  def finish_line
-    qualification_round.tournament.finish_line
-  end
-
-  def track_points
-    PointsQuery.execute(
-      track,
-      trimmed: { seconds_before_start: 20 },
-      only: %i[gps_time altitude latitude longitude]
-    )
   end
 end
